feat: Refactor authentication and admin pages to fix warning when build
