+++ LuxStudio/portalv2/src/app/(app)/page.tsx
@@ -1,4 +1,3 @@
-"use client";
 import { Stat } from '@/app/stat'
 import { Avatar } from '@/components/avatar'
 import { Heading, Subheading } from '@/components/heading'
